funções de crud feitas na classe resouce e endpoints ja testados no postman

+++ src/main/java/com/projeto/api/entidades/Clientes.java
@@ -30,7 +30,7 @@ private static final long serialVersionUID = 1L;
 
 	 private String nome;
 	
-	 private Integer celular;
+	 private Long celular;
 	
 	 private String endereco;
